Divide parts of code into components in Payment page
